+++ tests/testthat/test-geo_address_lookup.R
@@ -10,12 +10,14 @@ test_that("Returning empty query", {
 })
 
 test_that("Data format", {
+  skip_if_api_server()
   skip_if_offline()
   expect_true(is.data.frame(geo_address_lookup(34633854, "W")))
   expect_false(inherits(geo_address_lookup(34633854, "W"), "sf")) # this is _not_ a _sf function
 })
 
 test_that("Checking query", {
+  skip_if_api_server()
   skip_if_offline()
   expect_equal(ncol(geo_address_lookup(34633854, "W")), 4)
   expect_gt(ncol(geo_address_lookup(34633854, "W", full_results = TRUE)), 4)