+++ tests/testthat/test-geo_address_lookup_sf.R
@@ -8,12 +8,14 @@ test_that("Returning empty query", {
 })
 
 test_that("Data format", {
+  skip_if_api_server()
   skip_if_offline()
   expect_true(is.data.frame(geo_address_lookup_sf(34633854, "W")))
   expect_s3_class(geo_address_lookup_sf(34633854, "W"), "sf")
 })
 
 test_that("Checking query", {
+  skip_if_api_server()
   skip_if_offline()
   expect_equal(ncol(geo_address_lookup_sf(34633854, "W")), 3)
   expect_gt(ncol(geo_address_lookup_sf(34633854, "W", full_results = TRUE)), 3)