+++ tests/testthat/test-geo_amenity_sf.R
@@ -10,7 +10,9 @@ test_that("Returning empty query", {
 })
 
 test_that("Data format", {
+  skip_if_api_server()
   skip_if_offline()
+  
   expect_true(is.data.frame(geo_amenity_sf(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
     c("pub", "restaurant"),
@@ -23,7 +25,9 @@ test_that("Data format", {
 
 
 test_that("Checking query", {
+  skip_if_api_server()
   skip_if_offline()
+
   expect_equal(ncol(geo_amenity_sf(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
     c("pub", "restaurant"),