+++ tests/testthat/test-geo_lite.R
@@ -10,6 +10,7 @@ test_that("Returning empty query", {
 })
 
 test_that("Data format", {
+  skip_if_api_server()
   skip_if_offline()
   expect_true(is.data.frame(geo_lite("Madrid")))
   expect_false(inherits(geo_lite("Madrid"), "sf")) # this is _not_ a _sf function
@@ -17,7 +18,9 @@ test_that("Data format", {
 
 
 test_that("Checking query", {
+  skip_if_api_server()
   skip_if_offline()
+  
   expect_equal(ncol(geo_lite(c("Madrid", "Barcelona"))), 4)
   expect_gt(ncol(geo_lite("Madrid", full_results = TRUE)), 4)
   expect_gt(nrow(geo_lite("Madrid",