+++ tests/testthat/test-geo_address_lookup_sf.R
@@ -1,6 +1,9 @@
 test_that("Returning empty query", {
-  obj <- expect_message(geo_address_lookup_sf("xbzbzbzoa aiaia", "R"))
+  expect_message(geo_address_lookup_sf("xbzbzbzoa aiaia", "R"))
   skip_if_api_server()
+  
+  obj <- geo_address_lookup_sf("xbzbzbzoa aiaia", "R")
+  
   expect_true(ncol(obj) == 1)
   expect_true(nrow(obj) == 1)