+++ tests/testthat/test-geo_amenity_sf.R
@@ -1,10 +1,15 @@
 test_that("Returning empty query", {
-  obj <- expect_message(geo_amenity_sf(
+  expect_message(geo_amenity_sf(
     bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
     amenity = "xbzbzbzoa aiaia"
   ))
 
   skip_if_api_server()
+  
+  obj <- geo_amenity_sf(
+    bbox = c(-1.1446, 41.5022, -0.4854, 41.8795),
+    amenity = "xbzbzbzoa aiaia"
+  )
   
   expect_true(nrow(obj) == 1)
   expect_true(ncol(obj) == 1)